CR: move `qr/` folder to `proc/qrcode/`
