+++ scantools/proc/qrcode/detector.py
